backend-flask/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-jwt-extended
python-dotenv
werkzeug
email-validator
flask-mail
scikit-learn
pypdf
